Move internal members to .c file as is conventional
The original driver defines all internal global variables and functions in the .h files alongside the external ones. They use a define to choose whether the header should declare variables as extern or declare and define them. This was valid but confusing. Now the headers only contain the things that are relevant to other files. Internal variables and functions are in the .c file.